first pass of edits
It's not totally clear to me who the "user"/reader is for this document. Collectors? Or all of the site's users, including artists, curators, and collectors? Once I understand that it'll be easier to edit this doc.
Also, the term "curator" is confusing, as I haven't seen it in our other documents. I suppose "style owner" is the same as "curator"?